Added cancel button and event handler. Separated into controller and view (to the best that I could), and deleted unecessary imports
